+++ Changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
  - ```userPermissions``` function not taking into account Group Permissions
  - ```userPermissions``` method throwing error when no submission or submission id is specified
+ - ```Formio``` tests always failing with timeout error instead of actual error
 
 ## 4.8.0-beta.5
 ### Fixed